migrated to new parking system

+++ src/commonMain/kotlin/nl.parkeerassistent.model/Visitor.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Visitor(
     val visitorId: Int,
-    val permitId: Int,
+    val permitId: Long,
     val license: String,
     val formattedLicense: String,
     val name: String? = null
